+++ src/test/java/org/folio/mr/controller/delegate/BatchRequestsServiceDelegateTest.java
@@ -19,10 +19,12 @@ import org.folio.flow.api.FlowEngine;
 import org.folio.mr.client.SettingsClient;
 import org.folio.mr.domain.dto.MediatedBatchRequestDetailsDto;
 import org.folio.mr.domain.dto.MediatedBatchRequestDto;
+import org.folio.mr.domain.dto.MediatedBatchRequestDtoItemRequestsStats;
 import org.folio.mr.domain.dto.MediatedBatchRequestPostDto;
 import org.folio.mr.domain.dto.MediatedBatchRequestsDto;
 import org.folio.mr.domain.entity.MediatedBatchRequest;
 import org.folio.mr.domain.entity.MediatedBatchRequestSplit;
+import org.folio.mr.domain.entity.projection.BatchRequestStatsImpl;
 import org.folio.mr.domain.mapper.MediatedBatchRequestMapper;
 import org.folio.mr.exception.MediatedBatchRequestValidationException;
 import org.folio.mr.service.MediatedBatchRequestFlowProvider;
@@ -96,9 +98,10 @@ class BatchRequestsServiceDelegateTest {
 
   @Test
   void createBatchRequest_negative_shouldFailWithDuplicateItemsError() {
-    var split = new MediatedBatchRequestSplit();
-    split.setItemId(UUID.randomUUID());
-    split.setPickupServicePointId(UUID.randomUUID());
+    var split = MediatedBatchRequestSplit.builder()
+      .itemId(UUID.randomUUID())
+      .pickupServicePointId(UUID.randomUUID())
+      .build();
     var postDto = new MediatedBatchRequestPostDto();
     var batchEntity = mock(MediatedBatchRequest.class);
     var batchSplits = List.of(split, split);
@@ -111,12 +114,14 @@ class BatchRequestsServiceDelegateTest {
 
   @Test
   void createBatchRequest_negative_shouldFailWithItemsCountExceedLimitError() {
-    var split1 = new MediatedBatchRequestSplit();
-    split1.setItemId(UUID.randomUUID());
-    split1.setPickupServicePointId(UUID.randomUUID());
-    var split2 = new MediatedBatchRequestSplit();
-    split2.setItemId(UUID.randomUUID());
-    split2.setPickupServicePointId(UUID.randomUUID());
+    var split1 = MediatedBatchRequestSplit.builder()
+      .itemId(UUID.randomUUID())
+      .pickupServicePointId(UUID.randomUUID())
+      .build();
+    var split2 = MediatedBatchRequestSplit.builder()
+      .itemId(UUID.randomUUID())
+      .pickupServicePointId(UUID.randomUUID())
+      .build();
     var postDto = new MediatedBatchRequestPostDto();
     var batchEntity = mock(MediatedBatchRequest.class);
     var batchSplits = List.of(split1, split2);
@@ -135,8 +140,14 @@ class BatchRequestsServiceDelegateTest {
   void getBatchRequestById_positive_shouldReturnDto() {
     var id = UUID.randomUUID();
     var entity = mock(MediatedBatchRequest.class);
-    var expectedDto = mock(MediatedBatchRequestDto.class);
+    var expectedDto = new MediatedBatchRequestDto()
+      .itemRequestsStats(new MediatedBatchRequestDtoItemRequestsStats().total(10).completed(5).failed(5));
+    var stats = new BatchRequestStatsImpl();
+    stats.setTotal(10);
+    stats.setCompleted(5);
+    stats.setFailed(5);
     when(batchRequestsService.getById(id)).thenReturn(entity);
+    when(requestSplitService.getBatchRequestStats(id)).thenReturn(stats);
     when(mapper.toDto(entity)).thenReturn(expectedDto);
 
     var result = delegate.getBatchRequestById(id);